pyproject.toml
[tool.poetry]
name = "trip-planner-agent"
version = "0.1.0"
description = "A simple agent that plans trips for me"
authors = ["Vishnu Narayanan"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = {extras = ["tools"], version = "^0.22.5"}
pydantic = "^2.6.3"
unstructured = "==0.10.25"
langchain-groq = "^0.0.1"
python-dotenv = "1.0.0"
load-dotenv = "0.1.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
